corrected mistake in V19__create_product_table.sql

+++ coffee/src/main/resources/db/migration/V19__create_product_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE product (
-  id MEDIUMINT NOT NULL AUTO_INCREMENT, 
-  code VARCHAR (20),
-  name VARCHAR (50),
-  brand VARCHAR (50),
+  id INT NOT NULL AUTO_INCREMENT, 
+  code VARCHAR(20),
+  name VARCHAR(50),
+  brand VARCHAR(50),
   description VARCHAR(255),
   unit_price DECIMAL(10,2),
   quantity INT,
@@ -14,5 +14,5 @@ CREATE TABLE product (
   
   CONSTRAINT pk_product_id PRIMARY KEY (id),
   CONSTRAINT fk_product_category_id FOREIGN KEY (category_id) REFERENCES category (id),
-  CONSTRAINT fk_product_supplier_id FOREIGN KEY (supplier_id) REFERENCES user_detail (id)
-) ;
+  CONSTRAINT fk_product_supplier_id FOREIGN KEY (supplier_id) REFERENCES user_detail(id)
+);
